数值系统基础款完成 但暂未考虑建筑物类型
下面，完成建筑物级别抽象
给角色添加多种攻击盒子
修复角色旋转，攻击盒无法旋转bug

+++ Scripts/UI/HpBar.cs
@@ -16,7 +16,7 @@ public partial class HpBar : Control
     private Node2D _target = default!;
     private float _yOffset;
 
-    public void Initialize(Node2D target, int current, int max)
+    public void Initialize(Node2D target, float current, float max)
     {
         _target = target;
         // 在初始化时计算高度偏移
@@ -64,7 +64,7 @@ public partial class HpBar : Control
         return 60.0f;
     }
 
-    public void UpdateHealth(int current, int max)
+    public void UpdateHealth(float current, float max)
     {
         ProgressBar.MaxValue = max;
         var tween = CreateTween();
